move sweep to top level
